Adding very basic cache capability
Adding some very basic cache capability using a cache library from Kevin Dockx.  In general everything is publicly cached for 60 seconds, get Recipes/Brews/Water Profiles are not cached as these are not affected when a new item is added/removed/edited.  The Brew controller is cached privately as that will become user dependent anyway.

This is an example of the concept of caching rather than a full blown implementation.